==> 4_poi.ipynb <==
import skmob
import pandas as pd
import csv, os, sys, json
import urllib.request
from skmob.preprocessing import clustering
from pandas import Series, DataFrame
import math

device_id_list = []
for it in os.scandir('./staypoint'):
    if it.is_dir():
        if it.name.isdigit():
            device_id_list.append(it.name)
device_id_list.sort()

notPoiList = ['공공,사회기관', '시설,건물', '주택']
poidf = DataFrame()
for device_id in device_id_list:
    df = pd.read_csv('./cluster/'+device_id+'/category.csv')
    
    poiDict = dict()
    poiList = list()
    for idx in range(len(df)):
        category = df.loc[idx]['category']
        if category == 'none' or type(category) == float:
            continue

        category_list = category.split('>')

        category1 = category_list[0]


        if category1 not in notPoiList:
            poiDict[category] = 1
            poiList.append(category)
            if len(poiList) >= 3:
                break
    path = "./poi/"+device_id
    if not os.path.isdir(path):
        os.mkdir(path)
    
    # print(f"deviceid: {device_id}, poi: {poiList}")
    df = pd.DataFrame(poiList, columns=['poi'])
    df['deviceid'] = device_id
    df = df[['deviceid', 'poi']]
    # print(df)
    df.to_csv('./poi/' + device_id +  '/poi.csv', index=False)
    poidf = pd.concat([poidf, df])
# print(poidf)

poidf = pd.DataFrame()

total_dict = dict()
poiDict = dict()

# notPoiList = ['공공,사회기관', '시설,건물', '주택']

for device_id in device_id_list:
    df = pd.read_csv('./cluster/'+device_id+'/category.csv')

    inv_dict = dict()
    for idx in range(len(df)):
        cate = df.loc[idx]['category']
        if cate == 'none' or type(cate) == float:
            continue
        # print(f"category: {cate}, deviceid: {df.loc[idx]['deviceid']}")
        # if cate == 'none' or cate
        # print()
        # detail = cate.split('>')[1]
        category_list = cate.split('>')

        cate = category_list[0]

        if cate not in total_dict:
            total_dict[cate] = 1
        else:
            total_dict[cate] += 1


sorted_dict = sorted(total_dict.items(),
                     key=lambda item: item[1], reverse=True)
poidf = pd.DataFrame(sorted_dict, columns=['poi', 'count'])
poidf.to_csv('./poi.csv', index=False)
